+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog records changes to stable releases since 1.50.2. "TBA" changes he
 
 - fix: use default NVM directory if NVM_DIR is not set ([ref](https://github.com/microsoft/vscode/issues/133521))
 - fix: lines offset when debugging web worker extensions ([ref](https://github.com/microsoft/vscode/issues/136242))
+- fix: "copy as expression" and "add to watch" for private fields ([ref](https://github.com/microsoft/vscode/issues/135944))
 
 ## v1.62 (October 2021)